refactored code

+++ src/app/views/criar-usuario/criar-usuario.component.ts
@@ -3,15 +3,8 @@ import { Router } from '@angular/router';
 import { UsuarioService } from 'src/app/services/usuario.service';
 import { IUser } from 'src/app/shared/models/user.model';
 
-import { FormBuilder, FormControl, FormGroup, FormGroupDirective, NgForm, Validators } from  '@angular/forms' ;
-import { ErrorStateMatcher } from  '@angular/material/core' ;
+import { FormBuilder, FormGroup,  Validators } from  '@angular/forms' ;
 
-export class MyErrorStateMatcher implements ErrorStateMatcher {
-  isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
-    const isSubmitted = form && form.submitted;
-    return !!(control && control.invalid && (control.dirty || control.touched || isSubmitted));
-  }
-}
 
 @Component({
   selector: 'app-criar-usuario',
@@ -19,8 +12,6 @@ export class MyErrorStateMatcher implements ErrorStateMatcher {
   styleUrls: ['./criar-usuario.component.scss'],
 })
 export class CriarUsuarioComponent implements OnInit {
-  emailFormControl = new FormControl('', [Validators.required, Validators.email]);
-  matcher = new MyErrorStateMatcher();
 
   formulario: FormGroup;
   selectedFile: File | null = null;
@@ -34,11 +25,11 @@ export class CriarUsuarioComponent implements OnInit {
 
   ngOnInit(): void {
     this.formulario = this.formBuilder.group({
-      firstName: [''],
-      lastName: [''],
+      firstName: ['', Validators.required],
+      lastName: ['', Validators.required],
       title: [''],
       gender: [''],
-      email: [''],
+      email: ['', [Validators.required, Validators.email]],
       picture: [''],
       phone: [''],
       city: [''],
@@ -46,32 +37,24 @@ export class CriarUsuarioComponent implements OnInit {
     })
   }
 
-  onFileSelected(event: any) {
-    this.selectedFile = event.target.files[0];
-  }
-
   adicionarUsuario() {
-    const user: IUser = {
-      firstName: this.formulario.get('firstName').value,
-      lastName: this.formulario.get('lastName').value,
-      title: this.formulario.get('title').value,
-      gender: this.formulario.get('gender').value,
-      email: this.formulario.get('email').value,
-      picture: this.formulario.get('picture').value,
-      phone: this.formulario.get('phone').value,
-      location: {
-        city: this.formulario.get('city').value,
-        state: this.formulario.get('state').value,
-      }
-    };
+    if (this.formulario.invalid) {
+      console.error('Preencha os campos obrigatórios.');
+      return;
+    }
+
+    const user: IUser = this.formulario.value;
+    console.log('Usuário a ser criado: ', user);
 
     this.usuarioService.createUser(user).subscribe(
       (response) => {
+        console.log(response, 'chegou aqui')
         console.log('Usuário adicionado com sucesso: ', response);
         localStorage.clear();
         this.pagExibirTodos();
       },
       (error) => {
+        console.log('aqui chega', user)
         console.log('Erro ao tentar adicionar usuário: ', error);
       }
     );
